configure alembic and PosgreSQL
