--- SpendingTracker/internal/config/config.yaml
@@ -1,7 +1,0 @@
-log_level: INFO
-database:
-  user: user
-  password: password
-  host: localhost
-  port: 5432
-  name: name
