+++ internal/config/config.yaml
@@ -1,0 +1,7 @@
+log_level: INFO
+database:
+  username: postgres
+  password: password123
+  host: localhost
+  port: 5432
+  name: SpendingTrackerDB
